run_make_support: move `build_native_static_lib` under `external_deps`
And fix some `ar`-related rebase mishaps.
