+++ src/tools/run-make-support/src/external_deps/mod.rs
@@ -6,6 +6,7 @@
 //! - This is not the *only* place where external dependencies are assumed or referenced. For
 //!   example, see [`cygpath_windows`][crate::path_helpers::cygpath_windows].
 
+pub mod c_build;
 pub mod cc;
 pub mod clang;
 pub mod htmldocck;
